+++ CHANGES.md
@@ -7,6 +7,9 @@ Modificaciones post produccion
 *Cambios:*
     - Se cambia la visualización de housing por la cantidad de servicios que posea el cleinte
     M[pages/cdc_cliente.php]
+    - Se cambia el popup para visualizar grilla con la cantidad de servicios de housing.
+    M[pages/modals/sdc_housing_view.js]
+    M[pages/modals/sdc_housing_view.php]
 
 ## FEAT-CLI-DC